+++ src/pages/shop/components/ShopHeader.tsx
@@ -1,5 +1,5 @@
+import { Header } from '@/components/common';
 import { Modal } from '@jae-o/modal-component-module';
-import Header from '../../../../components/common/Header';
 import CartButton from './CartButton';
 
 function ShopHeader({ itemsCount }: { itemsCount: number }) {
