+++ src/pages/shop/components/index.ts
@@ -1,0 +1,3 @@
+export { default as ShopFilter } from './ShopFilter';
+export { default as ShopHeader } from './ShopHeader';
+export { default as ShopProductList } from './ShopProductList';
